// frontend/src/types/elements.ts
export const enum ElementNames {
  TEXTBOX = 'textbox',
  TEXT = 'text',
  ITEXT = 'i-text',
  INPUTTEXT = 'itext',
  ARCTEXT = 'arctext',
  VERTICALTEXT = 'verticaltext',
  IMAGE = 'image',
  SVGIMAGE = 'svgimage',
  CROPIMAGE = 'cropimage',
  MASK = 'mask',
  PATH = 'path',
  RECT = 'rect',
  LINE = 'line',
  ARROW = 'arrow',
  POLYLINE = 'polyline',
  ELLIPSE = 'ellipse',
  QRCODE = 'qrcode',
  BARCODE = 'barcode',
  GROUP = 'group',
  ACTIVE = 'activeselection',
  CIRCLE = 'circle',
  REFERENCELINE = 'referenceline',
  DOT = 'dot',
}
// ...
export const enum AlignCommand {
  LEFT = 'left',
  RIGHT = 'right',
  HORIZONTAL = 'horizontal',
  VERTICAL = 'vertical',
  CENTER = 'center',
  TOP = 'top',
  BOTTOM = 'bottom',
}

export const enum LayerCommand {
  UP = 'left',
  DOWN = 'right',
  TOP = 'top',
  BOTTOM = 'bottom',
}